// src/enum/config.enum.ts
export enum ConfigEnum {
  // MYSQL
  DB_TYPE = 'DB_TYPE',
  DB_HOST = 'DB_HOST',
  DB_PORT = 'DB_PORT',
  DB_DATABASE = 'DB_DATABASE',
  DB_USERNAME = 'DB_USERNAME',
  DB_PASSWORD = 'DB_PASSWORD',
  DB_SYNC = 'DB_SYNC',
  // MONGO
  MONGO_DB_HOST = 'MONGO_DB_HOST',
  MONGO_DB_PORT = 'MONGO_DB_PORT',
  MONGO_DB_DATABASE = 'MONGO_DB_DATABASE',
  MONGO_DB_USERNAME = 'MONGO_DB_USERNAME',
  MONGO_DB_PASSWORD = 'MONGO_DB_PASSWORD',
  // JWT
  SECRET = 'SECRET',
}

export enum LogEnum {
  LOG_LEVEL = 'LOG_LEVEL',
  LOG_ON = 'LOG_ON',
  TIMESTAMP = 'TIMESTAMP',
}